improve fix for directory checkout error
In #63, commands were added to remove the LICENSE.chromedriver file, which was
reportedly created by calling `chromedriver --version`.

It seems the license file is (now, at least) created when extracting the
downloaded zipfile.

This changeset skips the removal step and instead instructs the unzip command
to only extract the single file we want: `chromedriver`.